Refactorizar arquitectura basre
