+++ src/app/login/auth-success.vue
